fix(parser): frozen graph shouldn't block selection
